sign_language_api/requirements.txt
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.3.1
google-auth==2.1.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.26.1
google-cloud-translate==3.0.2

